Renamed ordered node graph to indexed node graph
