+++ src/app/graph2/inputhandler/IndexedGraphStartMarkerInputHandler.js
