+++ src/app/modules/fsa2/graph/FSAGraphController.js
@@ -2,7 +2,7 @@ import GraphController from 'graph2/controller/GraphController.js';
 
 import { GRAPH_EVENT_NODE_EDIT_WHILE_DELETE, GRAPH_EVENT_NODE_DELETE, GRAPH_EVENT_NODE_DELETE_ALL } from 'graph2/inputhandler/GraphNodeInputHandler.js';
 import { GRAPH_EVENT_EDGE_EDIT_WHILE_DELETE, GRAPH_EVENT_EDGE_DELETE } from 'graph2/inputhandler/GraphEdgeInputHandler.js';
-import { GRAPH_EVENT_START_MARKER_CHANGE } from 'graph2/inputhandler/GraphStartMarkerInputHandler.js';
+import { GRAPH_EVENT_START_MARKER_CHANGE } from 'graph2/inputhandler/IndexedGraphStartMarkerInputHandler.js';
 
 import { WARNING_LAYOUT_ID } from 'session/manager/notification/NotificationManager.js';
 import FSAGraphLabeler from './FSAGraphLabeler';
